Fixes / Changes / New Stuff - Version 0.9.5
Version 0.9.5

CHANGES:
- Revamped the Grid-Map / spawn point system
- Spawn Points don't use blocks any more they are now point and click
for the position (click the button in the inspector and click points on
the scene).
- Added in GUI button for placing in spawn points
- Added in visual identity to show where each spawn point is at (Cyan =
player / yellow = ai)
- Added in manual entry for spawn points (use at your own risk, this
method doesn't check for out of bounds area)
- Fixed up blocks not being linked to the parent prefab (before whenever
you changed the prefab the blocks on the scene wouldn't change)
P.S. - The current mazes need to be remade again mainly due to the fact
that the blocks aren't linked to the parent prefab, and also you will
see
pink blocks on the scene these were the spawn blocks, which I've deleted
you will have the use the spawn placer button.
- this is going to be the last time I change anything associated with
the grid system now.

- New Sounds Have Been Added (Thanks ....).
- Renamed a shit load of classes
- Moved the Grid Controller Into Its Own Object instead of being
attached to the grid-map
- Changed Sound-Controller to play music off current gameState.
- Cleaned Up the Inspector GUI for the Grid-Map

FIXES:
- Fixed A little bug where the AI would Stop In The Air and not detect
the player if it was under it (might still happen)
- Whenever the Grid-Map is selected on the Hierarchy Tab the scene
camera will now focus onto the Grid-Map (You will not be able to move
around the scene if the Grid-Map is still selected, you can zoom-in and
out tho.)

ADDED:
- Added in place-holders for the main menu / added in main menu scene.
- most of the buttons should have a mouseOn and mouseExit State.
- Added in place-holders for In-game GUI.
- Added in move count place-holder.
- Added in reset button place-holder.

COMING:
- Going to move the reset button into a menu screen
- going to place a menu button on the top right of the screen.
- menu transitions.
- Need a Sky-box for the menu and in-game mazes
- Finish up the main menu to link to different mazes
- Placer-holders will be changed when I've made a template assets for
it.
- The multi-directional block
- Undo movement
- Save System.